Documented every backend class. Refactored and fixed stuff in process.
