+++ src/main/java/me/pr0crustes/backend/classes/PDFConverter.java
@@ -12,28 +12,35 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+/**
+ * Class that handles image to pdf convertion.
+ */
 public class PDFConverter {
 
     private final File[] fileArray;
 
+    /**
+     * Constructor that fills important stuff.
+     * @param fileArray the array of files that should be converted into a single pdf.
+     */
     public PDFConverter(File[] fileArray) {
-        System.setProperty("sun.java2d.cmm", "sun.java2d.cmm.kcms.KcmsServiceProvider");
         this.fileArray = Objects.requireNonNull(fileArray);
     }
 
+    /**
+     * Method that converts every file in fileArray and returns a single pdf with everyone.
+     * @return a single PDDocument made of all images.
+     * @throws NoFileException in case of file related errors.
+     * @throws ArgumentException in case of invalid arguments.
+     */
     public PDDocument getDocumentFromImages() throws NoFileException, ArgumentException {
-
-        PDDocument document = new PDDocument();
-
         List<BufferedImage> bufferedImages = new ArrayList<>();
 
         try {
-
             for (File file : this.fileArray) {
                 BufferedImage fileAsImage = ImageIO.read(file);
                 bufferedImages.add(fileAsImage);
             }
-
         } catch (IOException e) {
             throw new NoFileException();
         } catch (NullPointerException e) {
