+++ src/main/java/me/pr0crustes/backend/classes/PDFQualityModifier.java
@@ -11,16 +11,29 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Class that handles changing a pdf quality (dpi).
+ */
 public class PDFQualityModifier {
 
     private final File file;
 
+    /**
+     * Constructor that receives the file to be changed.
+     * @param file the File that will get the quality change.
+     */
     public PDFQualityModifier(File file) {
         this.file = file;
     }
 
+    /**
+     * Method that handles the dpi of the file.
+     * @param dpi the target dpi.
+     * @return a PDDocument with the new dpi.
+     * @throws NoFileException in case of file error.
+     * @throws ArgumentException in case of invalid arguments.
+     */
     public PDDocument getDocumentWithDPI(int dpi) throws NoFileException, ArgumentException {
-
         PDDocument originalDocument = PDFManager.getFileDocument(this.file);
         PDFRenderer renderer = new PDFRenderer(originalDocument);
 
@@ -40,4 +53,5 @@ public class PDFQualityModifier {
 
         return pdfCreator.getDocument();
     }
+
 }
