+++ src/main/java/me/pr0crustes/backend/exeptions/ArgumentException.java
@@ -1,7 +1,9 @@
 package me.pr0crustes.backend.exeptions;
 
+/**
+ * ArgumentException is a custom exception that should be thrown when any argument is completely wrong.
+ */
 public class ArgumentException extends Exception {
-
     public ArgumentException() {
         super();
     }
