+++ src/main/java/me/pr0crustes/backend/exeptions/NoFileException.java
@@ -2,8 +2,10 @@ package me.pr0crustes.backend.exeptions;
 
 import java.io.IOException;
 
+/**
+ * NoFileException is a custom exception that should be thrown when a file is not found.
+ */
 public class NoFileException extends IOException {
-
     public NoFileException() {
         super();
     }
